+++ NEWS.md
@@ -1,7 +1,19 @@
 # OpenDDS Releases
 
+## Version 3.15 of OpenDDS
+OpenDDS 3.15 is currently in development, so this list might change.
+
+### Additions:
+- TODO: Add your features here
+
+### Fixes:
+- TODO: Add your fixes here
+
+### Notes:
+- TODO: Add your notes here
+
 ## Version 3.14 of OpenDDS
-OpenDDS 3.14 is currently in development, so this list might change.
+OpenDDS 3.14 was released on Mar  5 2020.
 
 ### Additions:
 - XTypes IDL annotations for topic types, Union topic types (#1067, #1422):
@@ -49,6 +61,7 @@ OpenDDS 3.14 is currently in development, so this list might change.
 - GuardCondition _narrow (#1407)
 - "make install" with DESTDIR (#1429, #1440)
 - Support for multiple transport instances in the same transport configuration (#1465)
+- The reliability QoS for DataWriters and Topics is now immutable, matching the DDS 1.4 spec (#1584)
 
 ### Notes:
 - The configure script now defaults to not including tests (#1287)